add scroll on bottom sheet

+++ app/src/main/java/br/com/reconecta/api/model/auth/LoginResponse.kt
@@ -37,6 +37,6 @@ data class Address(
     val city: String,
     val state: String,
     val postalCode: String,
-    val latitude: Int? = null,
-    val longitude: Int? = null,
+    val latitude: Float? = null,
+    val longitude: Float? = null,
 )
